requirements.txt
langchain>=0.1.0
pypdf>=3.0.0
python-docx>=0.8.11
sentence-transformers>=2.2.2
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.10.0
python-multipart>=0.0.5
numpy>=1.21.0
faiss-cpu>=1.7.0
